♻️ XPCNRE code: split cli harness into focused modules
