+++ .agentplane/tasks/202604172036-XPCNRE/pr/github-title.txt
@@ -1,0 +1,1 @@
+refactor/testkit: Decompose CLI test harness into focused fixture modules (XPCNRE)
