+++ .agentplane/tasks/202604172036-XPCNRE/pr/review.md
@@ -1,0 +1,73 @@
+# PR Review
+
+Created: 2026-04-17T20:56:34.564Z
+Branch: task/202604172036-XPCNRE/cli-harness-modules
+
+## Summary
+
+Decompose CLI test harness into focused fixture modules
+
+Split the monolithic CLI harness into focused fixture modules for git repo setup, stdio capture, recipe fixtures, and command execution while preserving the @agentplane/testkit facade and existing test behavior.
+
+## Scope
+
+- In scope: Split the monolithic CLI harness into focused fixture modules for git repo setup, stdio capture, recipe fixtures, and command execution while preserving the @agentplane/testkit facade and existing test behavior.
+- Out of scope: unrelated refactors not required for "Decompose CLI test harness into focused fixture modules".
+
+## Verification
+
+### Plan
+
+1. Run `bun run lint:core`. Expected: it succeeds and confirms the requested outcome for this task.
+2. Run `bun run test:fast`. Expected: it succeeds and confirms the requested outcome for this task.
+3. Review the changed artifact or behavior for the `code` task. Expected: the requested outcome is visible and matches the approved scope.
+4. Compare the final result against the task summary and touched scope. Expected: remaining follow-up is either resolved or explicit in ## Findings.
+
+### Current Status
+
+- State: ok
+- Note: Command: bun run framework:dev:bootstrap
+Result: pass; repo-local runtime rebuilt for the worktree after moving testing modules.
+
+Command: bun run lint:core
+Result: pass.
+
+Command: bun run typecheck
+Result: pass.
+
+Command: bunx vitest run packages/testkit/src/index.test.ts packages/testkit/src/cli.test.ts packages/agentplane/src/runner/usecases/scenario-materialize-task.test.ts packages/agentplane/src/runner/usecases/task-run-lifecycle.test.ts packages/agentplane/src/cli/run-cli.core.tasks.scaffold-derive.test.ts
+Result: pass (5 files, 15 tests).
+
+Command: bun run test:fast
+Result: pass (211 files, 1269 tests, 2 skipped).
+
+Observation: the first full test:fast run exposed a regression in createUpgradeBundle after the module move; the manifest path was corrected in the same task and the failing upgrade tests plus the full suite were re-run successfully.
+
+## Risks
+
+- Risk level: not recorded
+- Breaking change: no
+
+### Rollback
+
+- Revert task-related commit(s).
+- Re-run required checks to confirm rollback safety.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-04-17T20:56:34.564Z
+- Branch: task/202604172036-XPCNRE/cli-harness-modules
+- Head: 94366eade38f
+
+```text
+No changes detected.
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
